pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-FlakeForge"
version = "0.1.0"
description = "Flakeforge environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.8.0",
    "pytest>=8.3.0",
    "libcst>=1.4.0",
    "tree-sitter>=0.22.0",
    "tree-sitter-python>=0.21.0",
    "tree-sitter-javascript>=0.21.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m FlakeForge.server.app
server = "FlakeForge.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["FlakeForge", "FlakeForge.server"]
package-dir = { "FlakeForge" = ".", "FlakeForge.server" = "server" }